Templates

BeaconK8sUtilities currently supports one template: a template for creating scripts to setup tensorboard pods to follow the logs of a project.

BeaconK8sUtilities.setup_follow — Method

setup_follow(destination::AbstractString;
             namespace::AbstractString=get_current_namespace(), overwrite=false)

Sets up a follow.sh script in destination directory destination, which does not have to exist ahead of time.

namespace: name of the Kubernetes namespace to use. Defaults to the current active one introspected by get_current_namespace.

overwrite: whether or not to overwrite existing configuration scripts.

Creates:

$destination/follow.sh: shell script to wait for a pod to be ready (with progress information) and then follow its logs.

BeaconK8sUtilities.setup_tensorboard — Method

setup_tensorboard(destination::AbstractString; app::AbstractString,
                  logdir::AbstractString,
                  ecr::AbstractString = default_ecr(),
                  service_account::AbstractString = default_service_account(),
                  namespace::AbstractString = get_current_namespace(),
                  local_port::Int=6006, overwrite=false)

Sets up a tensorboard launch scripts in destination directory destination, which does not have to exist ahead of time.

app: name of project/application/model that you are using tensorboard with (used for labelling the pods and keeping track of which is which)

logdir: log directory that tensorboard should be pointing at. May be an s3 URI

ecr: name of the ECR repo to use to push docker images to. Can set this as a default with default_ecr.

service_account: name of the Kubernetes service account. Can set this as a default with default_service_account.

namespace: name of the Kubernetes namespace to use. Defaults to the current active one introspected by get_current_namespace.

local_port: local port to use when port forwarding.

overwrite: whether or not to overwrite existing configuration scripts.

Creates:

$destination/tensorboard.sh: shell script that either connects to an existing tensorboard pod or builds and launches a new tensorboard pod.

$destination/tensorboard.dockerfile: dockerfile to build tensorboard with access to s3 log directories (used by tensorboard.sh)

$destination/tensorboard.yaml: pod spec used to launch pods ((used by tensorboard.sh)

These are pre-populated using the keyword arguments passed here.