Utilities

As the name suggests, BeaconK8sUtilities provides several utilities for Julia-side management of Kubernetes pods, mostly intended for writing interactive scripts (such as those generated by our Templates).

BeaconK8sUtilities.follow — Method

follow(pod; exit_on_interrupt=false, namespace=nothing)

A simpler helper to follow a pod by first waiting until it's ready (wait_until_pod_ready) and then wathing its logs (watch_logs).

BeaconK8sUtilities.get_current_namespace — Method

get_current_namespace() -> String

Get the current active Kubernetes namespace.

BeaconK8sUtilities.get_pod_names — Function

get_pod_names(labels = ``; namespace=nothing) -> Vector{String}

Get the names of pods in the current namespace (or pass a namespace to specify a different one). Pass labels as e.g. -l app=myapp to restrict to specific labels.

BeaconK8sUtilities.get_status — Method

get_status(pod; namespace=nothing) -> String

Get the status of a pod.

BeaconK8sUtilities.last_condition — Method

last_condition(pod; namespace=nothing) -> String

Get the latest "condition" of a pod.

BeaconK8sUtilities.port_forward — Method

port_forward(pod; remote_port::Int, local_port::Int=remote_port, namespace=nothing) -> Task

Forwards a port from remote_port on pod to local_port. Runs in a detached process, and so will outlive the current Julia session. Returns the task running this process.

BeaconK8sUtilities.wait_until_pod_ready — Method

wait_until_pod_ready(pod; namespace=nothing)

Display a progress bar while waiting for pod to become Running.

BeaconK8sUtilities.watch_logs — Method

watch_logs(pod; exit_on_interrupt=false, namespace=nothing)

Follows logs from pod. If exit_on_interrupt=true, exits the Julia session upon ctrl-c. Otherwise throws an error as usual upon interruption.